requirements-dev.txt
-r requirements.txt
flake8
pytest
django-extensions
pytest-django
yapf
